Add function to get playcard suit as string
"playcard_suit_as_string" returns a string, actually a value of
playcard_suit_enum, indicating the suit of the card, extracted from the
second char of the playcard parameter passed to the function. The output
may be 'hearts', 'diamonds', 'clubs' or 'spades' for a known suit or
simply NULL for jokers and card backs.