Add function to get the playcard color
"playcard_suit_color" returns a string, actually a value of
playcard_color_enum, indicating the color of the suit of the card,
extracted from the second char of the playcard parameter passed to the
function. The output may be 'red' (for hearts and diamonds) or 'black'
(for clubs and spades). Jokers have also a color, even if they don't
have a suit. The color is picked from the suit they are assigned to,
e.g.: 'SH' would be a joker of hearts, but is actually a red joker.
Card backs produce a NULL output.