Add function to produce a playcard from a human readable description
"playcard_from_full_name" returns a playcard variable from parsing an
english human readable explanation of the card. It is the inverse
function of "playcard_full_name". The parsed string should contain the
value and the suit in this order, with optional punctuation and an 'of'
word. Jokers should have a color and card backs just the word 'back'.
A few examples:

- 'ace of spades' --> 'AS'
- 'KING    diamONDs' --> 'KD'
- '3 clubs' or 'three clubs' or 'three of clubs' --> '3C'
- 'clubs 3' --> Error!
- 'joker red' or 'red ...:::somegarbage::;-     joker' -> 'SH'
- 'back' or 'card back' --> 'BC'