Add value symbol and use char/varchar instead of text
The value symbol is a char(2) that shows the value of a card in a
compact way. It's a string alternative to the value_smallint. The
difference is visible for:

- ace: 'A' instead of 1
- jack: 'J' instead of 11
- queen: 'Q' instead of 12
- king: 'K' instead of 13
- joker: '★' or '☆' instead of NULL since it has no value

This could be useful when rendering the actual card on screen: an ace of
spades becomes just 'A♠' which is value_symbol || suit_symbol.

*****

The usage of char and varchar instead of text is for SQL code
portability and to enforce the string length constraint. For PostgreSQL
char/varchar/text makes no difference under the hood, but other RDBMS
may operate differently.

So for now to make code portable, just remove the CONSTRAINT clauses,
convert enums to varchar(8) and remove the function is_empty().